refactored swe l1a code and added tests
